Add color for password and hint strings
Added a hash that will return a colored string("1"-"6"), and a hint hash
that will return one of two symbols, to indicate if something matches or
not. Since I will have to likely print the full password out, I have a
basic print method implemented. This currently just prints out the
code_colors method. I may update it for better variety in the future.