Add basic password logic
I added a way for the AI to generate a random 4 digit password, through
sampling a repeated_permutation. I then added a simple way for users to
enter a password for their guess, but have not used it to compare and
print if they match yet. As an addition to this, I also added a
multi-layer guard to assure that the user only entered numbers between
1-6, and for Quality of Life, while retaining user friendliness.