Add comparison logic and print outcome
I had to convert the hint from my function, since it was an array of
strings "?" and "!". I iterated over these using a method to
convert them into the hint colors/shapes. I then made the logic to
check the password and guess, and put them through the methods to
print out the current state of the game.