Split hint method into smaller chunks
I needed to split them to get the values I needed more easily.
As a lazy programmer, I created a third method with the original name
that just runs both those new methods, so I did not have to rename
everything in other parts of the code. If this becomes and issue, I
will change it.